requirements.txt
# Dev dependencies used to run and build WhisperDesk (end users of the .exe don't need these)

# GUI
PySide6>=6.7

# Transcription
faster-whisper>=1.0.0
ctranslate2>=4.4.0

# Diarization
onnxruntime>=1.16.0
webrtcvad>=2.0.10
scikit-learn>=1.4.0

# Audio/processing
numpy>=1.25
soundfile>=0.12

# Exporters
python-docx>=1.0.0

# System info
psutil>=5.9
pynvml>=11.0.0

# Build (Windows)
pyinstaller>=6.0
pyinstaller-hooks-contrib>=2024.0
